+++ Assets/Scripts/Application Layer/UnitSystem/Interface/IMoneyData.cs
@@ -2,6 +2,6 @@ using UnityEngine;
 
 public interface IMoneyData
 {
-    int money { get; }
+    long money { get; }
     int carrot { get; }
 }
